feat:Refactor routes, order & cart files
