--- server/model/newOrderModel.js
@@ -1,66 +1,0 @@
-const mongoose = require("mongoose");
-
-const newOrderSchema = new mongoose.Schema(
-    {
-        user: {
-            type: mongoose.Schema.Types.ObjectId,
-            ref: "user",
-            required: true,
-        },
-        addressId: {
-            type: mongoose.Schema.Types.ObjectId,
-            ref: "UserAddress.address",
-            required: true
-        },
-        totalAmount: {
-            type: Number,
-            required: true
-        },
-        items: [
-            {
-                productId: {
-                    type: mongoose.Schema.Types.ObjectId,
-                    ref: "product"
-                },
-                payablePrice: {
-                    type: Number,
-                    required: true,
-                },
-                purchasedQty: {
-                    type: String,
-                    required: true
-                },
-            },
-        ],
-        paymentStatus: {
-            type: String,
-            enum: ["pending", "completed", "cancelled", "refund"],
-            required: true,
-        },
-        paymentType: {
-            type: String,
-            enum: ["cod", "card"],//cod mean cash on delivery
-            required: true,
-        },
-        orderStatus: [
-            {
-                type: {
-                    type: String,
-                    enum: ["ordered", "packed", "shipped", "delivered"],
-                    default: "ordered",
-                },
-                date: {
-                    type: Date,
-                },
-                isCompleted: {
-                    type: Boolean,
-                    default: false,
-                },
-            },
-        ],
-    },
-    { timestamps: true }
-
-);
-
-module.exports = mongoose.model("neworder", newOrderSchema)
