+++ server/model/orderModel.js
@@ -1,22 +1,66 @@
-const mongoose = require('mongoose')
+const mongoose = require("mongoose");
 
-const orderSchema = new mongoose.Schema({
-    user: { type: mongoose.Schema.Types.ObjectId, ref: 'user', required: true },
-    orderItems: [
-        {
-            product: { type: mongoose.Schema.Types.ObjectId, ref: 'product', required: true },
-            quantity: { type: Number, default: 1 },
-            price: { type: Number, required: true }
-        }
-    ]
+const newOrderSchema = new mongoose.Schema(
+    {
+        user: {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: "user",
+            required: true,
+        },
+        addressId: {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: "UserAddress.address",
+            required: true
+        },
+        totalAmount: {
+            type: Number,
+            required: true
+        },
+        items: [
+            {
+                productId: {
+                    type: mongoose.Schema.Types.ObjectId,
+                    ref: "product"
+                },
+                payablePrice: {
+                    type: Number,
+                    required: true,
+                },
+                purchasedQty: {
+                    type: String,
+                    required: true
+                },
+            },
+        ],
+        paymentStatus: {
+            type: String,
+            enum: ["pending", "completed", "cancelled", "refund"],
+            required: true,
+        },
+        paymentType: {
+            type: String,
+            enum: ["cod", "card"],//cod mean cash on delivery
+            required: true,
+        },
+        orderStatus: [
+            {
+                type: {
+                    type: String,
+                    enum: ["ordered", "packed", "shipped", "delivered"],
+                    default: "ordered",
+                },
+                date: {
+                    type: Date,
+                },
+                isCompleted: {
+                    type: Boolean,
+                    default: false,
+                },
+            },
+        ],
+    },
+    { timestamps: true }
 
+);
 
-}, { timestamps: true, versionKey: false });
-
-orderSchema.set('toJSON', {
-    virtuals: true,
-    transform: function (doc, ret) { delete ret._id }
-});
-
-module.exports = mongoose.model("order", orderSchema);
-
+module.exports = mongoose.model("neworder", newOrderSchema)
